--- src/models/IUser.ts
@@ -1,5 +1,0 @@
-interface IUser {
-  id: string;
-  email: string;
-  password: string;
-}
